# Fix #1438: nullness checker performance pathology