--- a/framework/src/org/checkerframework/framework/source/SourceChecker.java
+++ b/framework/src/org/checkerframework/framework/source/SourceChecker.java
@@ -390,6 +390,9 @@ public abstract class SourceChecker extends AbstractTypeProcessor
     /** The enabled lint options */
     private Set<String> activeLints;
 
+    /** Supported options for this checker */
+    protected Set<String> supportedOptions;
+
     /**
      * The active options for this checker. This is a processed version of {@link
      * ProcessingEnvironment#getOptions()}: If the option is of the form "-ACheckerName@key=value"
@@ -1731,7 +1734,11 @@ public abstract class SourceChecker extends AbstractTypeProcessor
     @Override
     public final String getOption(String name, String def) {
 
-        if (!this.getSupportedOptions().contains(name)) {
+        if (supportedOptions == null) {
+            supportedOptions = this.getSupportedOptions();
+        }
+
+        if (!supportedOptions.contains(name)) {
             ErrorReporter.errorAbort("Illegal option: " + name);
         }