The core cause is that it determines invariant fields by computing declared types of all fields for every method invocation. I've fixed this to remember invariant fields in the store with related optimizations.

The changes in `SourceChecker` is another optimization. Invocation of `getSupportedOptions` could be heavy because it calls `Class#getCanonicalName` and `Class#getSimpleName` so many times and these methods are slow(See an attachment in [JDK-8187123](https://bugs.openjdk.java.net/browse/JDK-8187123) I'm not sure who should cache this. If `BaseTypeChecker` should do, please let me know.

Now the original source in the issue could be compiled in about 30 sec by `NullnessChecker`. I think this is acceptable. For another checker can compile it in about 10 sec like `ValueChecker`. `NullnessChecker` will bring about an execution of KeyForSubchecker.
So we can consider it takes about 15 sec averagely for each checker and this is similar to other checkers.

@wmdietl, in the previous mail, I wrote it's compiled in about 20 sec. However, it had a regression. It became about 30 sec when I fix the regression.