@wmdietl 2017-09-01
@bitterfox Will you have a chance to address these comments?