@bitterfox 2017-09-02
I've now fixed. Please check this again.