@bitterfox 2017-09-02
Now it can be compiled in about 20 sec - 30 sec.
```

[/home/bitter_fox/dev/gsoc/checker-framework]
bitter_fox@bitterfox-MacBookPro% time javac -processor NullnessChecker checker/jtreg/nullness/Issue1438.java                                                                             (git)-[fix_1438]  (m4):S44
Error message is omited
javac -processor NullnessChecker checker/jtreg/nullness/Issue1438.java  54.75s user 1.77s system 239% cpu 23.577 total

[/home/bitter_fox/dev/gsoc/checker-framework]
bitter_fox@bitterfox-MacBookPro% time javac -processor NullnessChecker checker/jtreg/nullness/Issue1438.java                                                                             (git)-[fix_1438]  (m4):S44
Error message is omited
javac -processor NullnessChecker checker/jtreg/nullness/Issue1438.java  72.18s user 1.30s system 257% cpu 28.558 total

[/home/bitter_fox/dev/gsoc/checker-framework]
bitter_fox@bitterfox-MacBookPro% time javac -processor NullnessChecker checker/jtreg/nullness/Issue1438.java                                                                             (git)-[fix_1438]  (m4):S44
Error message is omited
javac -processor NullnessChecker checker/jtreg/nullness/Issue1438.java  51.31s user 1.50s system 227% cpu 23.241 total                                    
```