@wmdietl 2017-09-04
I fixed my two small comments.